--- packages/ui/Scrollable/Scrollable.module.less
@@ -1,42 +1,0 @@
-/* Scrollable.module.less */
-.scrollableFill {
-	position: relative;
-	width: 100%;
-	height: 100%;
-}
-
-.scrollable {
-	.scrollableFill;
-
-	display: flex;
-	box-sizing: border-box;
-	flex-direction: column;
-
-	.container {
-		display: flex;
-		width: 100%;
-		height: 100%;
-		min-height: 0;
-		flex: auto;
-		[data-container-muted="true"]::before {
-			position: absolute;
-			top: 0;
-			right: 0;
-			bottom: 0;
-			left: 0;
-			content: '';
-			z-index: 10;
-		}
-	}
-
-	.content {
-		background: inherit;
-		position: relative;
-		overflow: hidden;
-		flex: 1 1 100%;
-	}
-
-	.contentNative {
-		-webkit-overflow-scrolling: touch;
-	}
-}
