--- packages/ui/Scroller/UiScroller.module.less
@@ -1,10 +1,0 @@
-// UiScroller.module.less
-//
-.scroller {
-	will-change: transform;
-
-	/* hide native scrollbar */
-	&::-webkit-scrollbar {
-		display: none;
-	}
-}
