--- packages/ui/VirtualList/UiVirtualList.module.less
@@ -1,31 +1,0 @@
-// UiVirtualList.module.less
-//
-.virtualList {
-	position: relative;
-	height: 100%;
-	width: 100%;
-	overflow: hidden;
-
-	&.vertical {
-		overflow-y: scroll;
-	}
-
-	&.horizontal {
-		overflow-x: scroll;
-	}
-
-	.listItem {
-		overflow: hidden;
-		width: 100%;
-
-		will-change: transform;
-	}
-
-	.content {
-		will-change: transform;
-	}
-}
-
-.virtualList::-webkit-scrollbar {
-	display: none;
-}
